internal/db/update: Move schema update helpers to update package
